# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для API заявок ремонтной мастерской",
        "contact": {
            "email": "cgptvlados117@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Для Логирования"
        }
    ],
    "paths": {
...
        "/api/sanctum/getUsersList": {
            "get": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Получение списка пользователей",
                "description": "Возвращает список всех зарегистрированных пользователей.",
                "operationId": "3a6ba1c4ddb8022f91045cd9fc0246d6",
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RegisterRequest"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован. Токен доступа отсутствует или недействителен."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/sanctum/getUsersByField": {
            "post": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Get users by field",
                "description": "Меняет данные пользователя по его Id или Email",
                "operationId": "b89417cabd19ff3c72f1b75f6a5460ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "role": {
                                        "description": "User role",
                                        "type": "string",
                                        "example": "master"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Users founded successfully"
                    },
                    "404": {
                        "description": "Users not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/sanctum/updateUserByIdOrEmail/{user}": {
            "put": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Update user by ID or email",
                "description": "Меняет данные пользователя по его Id или Email",
                "operationId": "56eee5ceaa78421f8eeb78aae208884b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID or email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "role": {
                                        "description": "User role",
                                        "type": "string",
                                        "example": "master"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/sanctum/getUserByIdOrEmail/{user}": {
            "get": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Get user by ID or email",
                "description": "Получает конкретного пользователя по ID или email",
                "operationId": "87512b69b78e1ffe0867617bc118d86b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID or email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User founded successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/sanctum/deleteUserByIdOrEmail/{user}": {
            "delete": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Delete user by ID or email",
                "description": "Удаляет конкретного пользователя по ID или email",
                "operationId": "edc3872a13bd1361f64f3cc7e6c051bd",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID or email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "RegisterRequest": {
                "title": "Регистрация нового пользователя",
                "required": [
                    "name",
                    "email",
                    "password",
                    "user_firstname",
                    "user_lastname",
                    "user_patronymic"
                ],
                "properties": {
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Ivan"
                    },
                    "email": {
                        "description": "Адрес электронной почты",
                        "type": "string",
                        "format": "email",
                        "example": "ivan@yandex.ru"
                    },
                    "password": {
                        "description": "Пароль пользователя (не менее 8 символов)",
                        "type": "string",
                        "format": "password",
                        "example": "12345678"
                    },
                    "user_firstname": {
                        "description": "Имя пользователя (не менее 3 символов)",
                        "type": "string",
                        "example": "Владислав"
                    },
                    "user_lastname": {
                        "description": "Фамилия пользователя (не менее 3 символов)",
                        "type": "string",
                        "example": "Остряков"
                    },
                    "user_patronymic": {
                        "description": "Отчество пользователя (не менее 3 символов)",
                        "type": "string",
                        "example": "Павлович"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    }
}